Добавляет кэширующие селекторы отзывов и править компоненты с русским языком
