+++ project/src/components/loading-failed/loading-failed.tsx
@@ -4,8 +4,8 @@ function LoadingFailed(): JSX.Element {
   return (
     <section className={styles['loading-failed']}>
       <div className={styles['loading-failed--info']}>
-        <h2 className={styles['loading-failed--info-title']}>Ошибка загрузки данных</h2>
-        <p className={styles['loading-failed--info-text']}>Повторите попытку позже</p>
+        <h2 className={styles['loading-failed--info-title']}>Data loading failed</h2>
+        <p className={styles['loading-failed--info-text']}>Try later</p>
       </div>
     </section>
   );
