+++ project/src/components/reviews/reviews.tsx
@@ -1,28 +1,11 @@
 import dayjs from 'dayjs';
 import { useSelector } from 'react-redux';
 import { AuthorizationStatus } from '../../const';
-import { getReviews } from '../../store/reviews/selectors';
+import { getReviews, selectSortedReviews } from '../../store/reviews/selectors';
 import { getAuthStatus } from '../../store/user/selectors';
 import { Review } from '../../types/review';
 import { convertRating } from '../../utils/common';
 import ReviewForm from '../review-form/review-form';
-
-const MAX_REVIEWS_RENDER = 10;
-
-const compareDates = (a: Review, b: Review) => {
-  const dateA = new Date(a.date);
-  const dateB = new Date(b.date);
-
-  if (dateA < dateB) {
-    return 1;
-  }
-
-  if (dateA > dateB) {
-    return -1;
-  }
-
-  return 0;
-};
 
 type ReviewsProps = {
   id: string;
@@ -31,15 +14,11 @@ type ReviewsProps = {
 function Reviews({id}: ReviewsProps): JSX.Element {
   const authStatus= useSelector(getAuthStatus);
   const reviewsFromBackend = useSelector(getReviews);
-
-  const sortedReviews = [...reviewsFromBackend].sort((a, b) => compareDates(a, b));
-  const reviews = sortedReviews.length > MAX_REVIEWS_RENDER
-    ? sortedReviews.splice(0, MAX_REVIEWS_RENDER)
-    : sortedReviews;
+  const reviews = useSelector(selectSortedReviews);
 
   return (
     <section className="property__reviews reviews">
-      <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
+      <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviewsFromBackend.length}</span></h2>
 
       <ul className="reviews__list">
         {reviews.map((review: Review) => {
